complete code refactor; finally separated routes into corresponding files and call them successfully from server.js
