+++ app/data/friends.txt
@@ -1,1 +1,1 @@
-[{"name":"ming","pic":"sunny_pic","resultsArr":["3","4","2","4","3","3","4","3","2","2"]},{"name":"jonny","pic":"car_pic","resultsArr":["3","3","1","1","2","3","4","2","4","3"]},{"name":"br","pic":"c","resultsArr":["2","5","5","5","5","5","5","5","5","5"]},{"name":"merry","pic":"asd","resultsArr":["4","1","3","1","3","4","3","2","5","3"]},{"name":"lone","pic":"jpeg.gpe","resultsArr":["3","4","4","3","3","3","2","3","2","3"]}]
+[{"name":"ming","pic":"sunny_pic","resultsArr":["3","4","2","4","3","3","4","3","2","2"]},{"name":"jonny","pic":"car_pic","resultsArr":["3","3","1","1","2","3","4","2","4","3"]},{"name":"br","pic":"c","resultsArr":["2","5","5","5","5","5","5","5","5","5"]},{"name":"merry","pic":"asd","resultsArr":["4","1","3","1","3","4","3","2","5","3"]},{"name":"lone","pic":"jpeg.gpe","resultsArr":["3","4","4","3","3","3","2","3","2","3"]},{"name":"micach","pic":"https://www.google.com.mx/search?q=micachu&source=lnms&tbm=isch&sa=X&ved=0ahUKEwj39M6tzcPeAhVH1oMKHZR-BqkQ_AUIDigB&biw=1280&bih=698#imgrc=JnBe4q8U5PwyeM:","resultsArr":["3","3","3","5","5","1","1","4","4","3"]},{"name":"frank","pic":"https://www.google.com.mx/search?biw=1280&bih=698&tbm=isch&sa=1&ei=7YvjW4jGHajojgS_s6rQDA&q=frank+oce&oq=frank+oce&gs_l=img.3..0l10.178793.180188.0.180516.9.8.0.1.1.0.131.797.5j3.8.0....0...1c.1.64.img..0.9.800...35i39k1.0.hUUQuYq3jC4#imgrc=zTlKt4OFcwZ2DM:","resultsArr":["3","2","3","2","5","1","3","4","5","2"]}]
